Rename folder and variable names for consistency with other code
