+++ src/mesh/boundary/nodal_load.hpp
@@ -1,7 +1,7 @@
 
 #pragma once
 
-#include "mesh/generic/boundary.hpp"
+#include "mesh/boundary/boundary.hpp"
 
 namespace neon
 {
