+++ src/mesh/diffusion/boundary/surface_boundary.hpp
@@ -1,7 +1,7 @@
 
 #pragma once
 
-#include "mesh/generic/neumann.hpp"
+#include "mesh/boundary/neumann.hpp"
 
 #include "newton_convection.hpp"
